[CHORE] #35 : move navigator common -> common/navigation
