+++ android/core/common/src/main/java/com/jaino/common/extensions/ContextExt.kt
@@ -1,0 +1,8 @@
+package com.jaino.common.extensions
+
+import android.content.Context
+import android.widget.Toast
+
+fun Context.showToast(message: String){
+    Toast.makeText(this, message, Toast.LENGTH_SHORT).show()
+}
