+++ android/core/common/src/main/java/com/jaino/common/navigation/AppNavigator.kt
@@ -1,9 +1,10 @@
-package com.jaino.common
+package com.jaino.common.navigation
 
 import android.content.Intent
 
 interface AppNavigator {
     fun navigateToAuth(): Intent
+
     fun navigateToSetting(): Intent
 
     fun navigateToAnalyze(): Intent
